requirements.txt
# Kwyre AI — inference server dependencies
torch==2.6.0
transformers==4.48.3
bitsandbytes==0.45.3
accelerate==1.3.0
psutil==7.1.0
huggingface_hub==0.28.1
safetensors==0.5.3
tokenizers==0.21.1
autoawq==0.2.8

# QAT training dependencies (not required for inference only)
peft==0.14.0
datasets==3.3.2
trl==0.15.2

# Utilities
cryptography==44.0.1
pyyaml==6.0.2
scipy==1.15.2
